reorganize-tests-folder-structure
